qdiff/requirements.txt
albumentations==0.4.3
diffusers==0.24.0
pudb==2019.2
invisible-watermark
imageio==2.9.0
imageio-ffmpeg==0.4.2
test-tube>=0.7.5
streamlit>=0.73.1
torch-fidelity==0.3.0
torchmetrics==0.6.0
streamlit-drawable-canvas==0.8
einops==0.3.0
kornia==0.6.9
lmdb==1.3.0
natsort==8.3.1
omegaconf==2.1.1
opencv_python
opencv_python_headless
pandas==1.4.2
Pillow==9.0.1
pytorch_lightning==1.4.2
PyYAML==6.0
six==1.16.0
tqdm==4.64.0
transformers==4.32.1
omegaconf

